LayoutTests/media/wireless-playback-media-player/current-time-updates-after-route-seek-expected.txt


Test that currentTime updates after a seek driven by the route's currentPlaybackPosition.

RUN(internals.mockMediaDeviceRouteController.activateRoute(route))
RUN(video.src = findMediaFile('video', '../content/test'))
RUN(route.playbackRate = 1)
RUN(route.timeRange = { start: 0, duration: 60 })
RUN(video.play())
RUN(route.ready = true)
EVENT(canplaythrough)
EVENT(timeupdate)
RUN(route.currentPlaybackPosition = 30)
EXPECTED (video.currentTime > 29) OK
END OF TEST
